perf: Loop over blocks only once
